Save product images to database
Added multiple functions for use in the /save-product endpoint to save product image urls to the firestore database as a array for use in the frontend.

The uploadProductImages function will loops through the images in the request body.

The uploadeImageToStorage function will upload the images first in the firebase storage to generate the urls.

The generateV4ReadSignedUrl generates the urls for the images and returns a promise for the image url used to create an array of image url promises

The generateSignedUrlArray creates the array of promises that will be resolved in the /save-product endpoint to be sent with the product data as an array of image urls to the firestore database.